+++ crates/pathfinder/src/bin/pathfinder/main.rs
@@ -28,8 +28,7 @@ async fn main() -> anyhow::Result<()> {
 
     setup_tracing();
 
-    let config =
-        config::Configuration::parse_cmd_line().context("Parsing configuration")?;
+    let config = config::Configuration::parse_cmd_line().context("Parsing configuration")?;
 
     info!(
         // this is expected to be $(last_git_tag)-$(commits_since)-$(commit_hash)
@@ -65,8 +64,8 @@ Hint: Make sure the provided ethereum.url and ethereum.password are good.",
 
     // Note that network testnet2 integration are mutually exclusive, which is already
     // checked in the config builder.
-    let network = match (config.network, config.testnet2, config.integration) {
-        (Some(network), _, _) => match network.as_str() {
+    let network = match config.network {
+        Some(network) => match network.as_str() {
             "mainnet" => Chain::Mainnet,
             "testnet" => Chain::Testnet,
             "testnet2" => Chain::Testnet2,
@@ -76,10 +75,8 @@ Hint: Make sure the provided ethereum.url and ethereum.password are good.",
                 anyhow::bail!("{other} is not a valid network selection. Please specify one of: mainnet, testnet, testnet2, integration or custom.")
             }
         },
-        (None, true, _) => Chain::Testnet2,
-        (None, _, true) => Chain::Integration,
         // Defaults if not specified
-        (None, _, _) => match ethereum_chain {
+        None => match ethereum_chain {
             EthereumChain::Mainnet => Chain::Mainnet,
             EthereumChain::Goerli => Chain::Testnet,
             EthereumChain::Other(id) => anyhow::bail!(